Modify CML and PGS equations in calc_net_energy_activity
CML and PGS have a single coefficient. It does not differ among low and high level activity. To avoid confusion with variable nomenclature and exception, the function was reviewed to ensure that the cact is now multiplying the sum of activity_fractio + high_activity_fraction for CML and PGS.
The test was also reviewed and updated